Refactor CLI architecture and improve baseline management robustness
- Extract GitHub API logic into `dev-tools/gh_client.py`.
- Move UI audit business logic into `dev-tools/audit_logic.py`.
- Implement robust JSON extraction in `dev-tools/utils.py` for cleaner inter-tool communication.
- Update `CONTRIBUTING.md` and `AGENTS.md` to prioritize GitHub Variables as the source of truth for technical debt.
- Refine CI `lockfile-integrity` job and ensure all quality gates are passing.
- Clean up redundant baseline files and artifacts.
