+++ scripts/detect-antipatterns.mjs
@@ -240,7 +240,7 @@ const totalViolations = Object.values(allViolations).flat().length;
 
 if (isJson) {
   process.stdout.write(JSON.stringify(allViolations, null, 2));
-  process.exit(totalViolations > 0 ? 1 : 0);
+  process.exit(0); // Exit 0 when JSON is requested so downstream can parse
 }
 
 // Always output violation count for CI consumption
